backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
psycopg[binary]
python-dotenv
